Wrap web server into main()
